refactor: Rename templates to renderer
