+++ internal/service/renderer/templates/welcome.gotmpl
