+++ internal/service/renderer/testdata/help.golden
