+++ internal/service/renderer/testdata/welcome.golden
